refactor: 重构 Connection Token 模块 (完成 Model 设计和创建 Token 的API逻辑)
