+++ apps/authentication/serializers/__init__.py
@@ -1,4 +1,5 @@
 from .token import *
-from .connect_token import *
+from .connect_token1 import *
+from .connection_token import *
 from .password_mfa import *
 from .confirm import *
