+++ apps/users/urls/api_urls.py
@@ -16,7 +16,7 @@ router.register(r'users', api.UserViewSet, 'user')
 router.register(r'groups', api.UserGroupViewSet, 'user-group')
 router.register(r'users-groups-relations', api.UserUserGroupRelationViewSet, 'users-groups-relation')
 router.register(r'service-account-registrations', api.ServiceAccountRegistrationViewSet, 'service-account-registration')
-router.register(r'connection-token', auth_api.UserConnectionTokenViewSet, 'connection-token')
+# router.register(r'connection-token', auth_api.UserConnectionTokenViewSet, 'connection-token')
 
 
 urlpatterns = [
